Move function into commands
